requirements.txt
injector==0.18.0
gpiozero==2.0.1
paho-mqtt==2.1.0
pydantic~=2.12.0
python-dotenv==1.0.1